+++ src/useQuery/useQuery.spec.tsx
@@ -7,9 +7,11 @@ import { act } from 'react-test-renderer';
 
 import useQuery from './useQuery';
 import generateQueryCacheKey from '../utils/generateQueryCacheKey';
-import { CACHE_QUERY_RESULT } from '../store/constants';
 import Context from '../Context';
 import CustomReduxContext from '../store/CustomReduxContext';
+import { queryRequest, querySuccess } from '../store/queries/actions';
+import { saveNormalizedEntities } from '../store/entities/actions';
+import { State } from '../store/reducer';
 
 const mockAPI = {
   users: {
@@ -17,9 +19,11 @@ const mockAPI = {
     list: options => {
       return new Promise((resolve, reject) => {
         if (options && options.page === 2) {
-          resolve('more data');
+          resolve({ data: [{ id: 'e57a6047-cb52-4273-9df7-1d55c2c6e36d' }] });
         }
-        resolve('data');
+        resolve({
+          data: [{ id: 'e6538393-aa7e-4ec2-870b-f75b3d85f706' }, { id: '708c8008-3455-49a9-b66a-5222bcadb0cc' }],
+        });
       });
     },
     info: options => {
@@ -30,13 +34,60 @@ const mockAPI = {
       });
     },
   },
+  projects: {
+    list: jest.fn()
+  },
+  projectItems: {
+    report: () => {
+      return new Promise((resolve) => {
+        resolve(
+          {
+            data: [
+              {
+                billable_amount: { "amount": 63.05, "currency": "EUR" },
+                cost: { "amount": 100, "currency": "EUR" },
+                result: { "amount": -36.95, "currency": "EUR" },
+                type: "tracked_time",
+                quantity: 1,
+                unit: "hour"
+              }
+            ]
+          }
+        )
+      })
+    }
+  }
+};
+
+const initialMockState: State = {
+  entities: {},
+  queries: {},
+};
+
+const mockedStateAfterUsersListResolve = {
+  queries: {
+    [generateQueryCacheKey({ domain: 'users', action: 'list' })]: {
+      loading: false,
+      ids: ['e6538393-aa7e-4ec2-870b-f75b3d85f706', '708c8008-3455-49a9-b66a-5222bcadb0cc'],
+    },
+  },
+  entities: {
+    users: {
+      'e6538393-aa7e-4ec2-870b-f75b3d85f706': {
+        id: 'e6538393-aa7e-4ec2-870b-f75b3d85f706',
+      },
+      '708c8008-3455-49a9-b66a-5222bcadb0cc': {
+        id: '708c8008-3455-49a9-b66a-5222bcadb0cc',
+      },
+    },
+  },
 };
 
 const StoreWrapper = (passedStore?: Store) => {
   let store = passedStore;
   // initialize a default store
   if (!passedStore) {
-    store = configureStore([])({ queries: {} });
+    store = configureStore([])(() => initialMockState);
   }
 
   return ({ children }) => (
@@ -63,7 +114,7 @@ describe('useQuery', () => {
     expect(result.current.fetchMore).toBeInstanceOf(Function);
   });
 
-  it('should dispatch a caching action when the resolved data is not cached yet', async () => {
+  it('should dispatch a queryRequest action when a query is requested for the first time', async () => {
     const QUERY = () => ({
       domain: 'users',
       action: 'list',
@@ -71,19 +122,55 @@ describe('useQuery', () => {
 
     const store = configureStore([])({ queries: {} });
 
-    const { waitForNextUpdate } = renderHook(() => useQuery(QUERY), { wrapper: StoreWrapper(store) });
+    renderHook(() => useQuery(QUERY), { wrapper: StoreWrapper(store) });
 
-    // wait for request to resolve
-    await waitForNextUpdate();
+    const key = generateQueryCacheKey({ domain: 'users', action: 'list' });
 
-    const expectedType = CACHE_QUERY_RESULT;
-    const expectedPayload = {
+    const action = queryRequest({ key });
+
+    expect(store.getActions()).toEqual([action]);
+  });
+
+  it('should dispatch a querySuccess and saveNormalizedEntities action when a query is successfully resolved', async () => {
+    const QUERY = () => ({
       domain: 'users',
       action: 'list',
-      data: 'data',
-    };
+    });
 
-    expect(store.getActions()).toEqual([{ type: expectedType, payload: expectedPayload }]);
+    const key = generateQueryCacheKey({ domain: 'users', action: 'list' });
+
+    let state = { ...initialMockState };
+    const store = configureStore([])(() => state);
+
+    const { waitForNextUpdate } = renderHook(() => useQuery(QUERY), { wrapper: StoreWrapper(store) });
+
+    state = mockedStateAfterUsersListResolve;
+
+    await waitForNextUpdate();
+
+    const actions = [
+      querySuccess({
+        key,
+        ids: ['e6538393-aa7e-4ec2-870b-f75b3d85f706', '708c8008-3455-49a9-b66a-5222bcadb0cc'],
+      }),
+      saveNormalizedEntities({
+        type: 'users',
+        entities: {
+          'e6538393-aa7e-4ec2-870b-f75b3d85f706': {
+            id: 'e6538393-aa7e-4ec2-870b-f75b3d85f706',
+          },
+          '708c8008-3455-49a9-b66a-5222bcadb0cc': {
+            id: '708c8008-3455-49a9-b66a-5222bcadb0cc',
+          },
+        },
+      }),
+    ];
+
+    const allActions = store.getActions();
+    // skip first request action
+    allActions.shift();
+
+    expect(allActions).toEqual(actions);
   });
 
   it('should return the resolved data', async () => {
@@ -92,13 +179,75 @@ describe('useQuery', () => {
       action: 'list',
     });
 
-    const { result, waitForNextUpdate } = renderHook(() => useQuery(QUERY), { wrapper: StoreWrapper() });
+    const key = generateQueryCacheKey({ domain: 'users', action: 'list' });
+
+    let state = { ...initialMockState };
+    const store = configureStore([])(() => state);
+
+    const { result, waitForNextUpdate } = renderHook(() => useQuery(QUERY), { wrapper: StoreWrapper(store) });
+
+    state = mockedStateAfterUsersListResolve;
 
     await waitForNextUpdate();
 
     expect(result.current.loading).toEqual(false);
-    expect(result.current.data).toEqual('data');
+    expect(result.current.data).toEqual([
+      {
+        id: 'e6538393-aa7e-4ec2-870b-f75b3d85f706',
+      },
+      {
+        id: '708c8008-3455-49a9-b66a-5222bcadb0cc',
+      },
+    ]);
   });
+
+  it('should save the data on a query level when the action is not "info" or "list"', async () => {
+    const QUERY = () => ({
+      domain: 'projectItems',
+      action: 'report'
+    });
+
+    const key = generateQueryCacheKey(QUERY());
+
+    let state = { ...initialMockState };
+    const store = configureStore([])(() => state);
+
+    const data = [{
+      billable_amount: { "amount": 63.05, "currency": "EUR" },
+      cost: { "amount": 100, "currency": "EUR" },
+      result: { "amount": -36.95, "currency": "EUR" },
+      type: "tracked_time",
+      quantity: 1,
+      unit: "hour"
+    }];
+
+    const { waitForNextUpdate } = renderHook(() => useQuery(QUERY), { wrapper: StoreWrapper(store) });
+
+    state = {
+      entities: {},
+      queries: {
+        [key]: {
+          loading: false,
+          data,
+        }
+      }
+    };
+
+    await waitForNextUpdate();
+
+    const actions = [
+      querySuccess({
+        key,
+        data
+      }),
+    ];
+
+    const allActions = store.getActions();
+    // skip first request action
+    allActions.shift();
+
+    expect(allActions).toEqual(actions);
+  })
 
   it('should return the error if the API request fails', async () => {
     const QUERY = () => ({
@@ -120,14 +269,19 @@ describe('useQuery', () => {
       action: 'list',
     });
 
-    const storeKey = generateQueryCacheKey(QUERY());
-
-    const store = configureStore([])({ queries: { [storeKey]: 'cachedData' } });
+    const store = configureStore([])(mockedStateAfterUsersListResolve);
 
     const { result } = renderHook(() => useQuery(QUERY), { wrapper: StoreWrapper(store) });
 
     expect(result.current.loading).toEqual(false);
-    expect(result.current.data).toEqual('cachedData');
+    expect(result.current.data).toEqual([
+      {
+        id: 'e6538393-aa7e-4ec2-870b-f75b3d85f706',
+      },
+      {
+        id: '708c8008-3455-49a9-b66a-5222bcadb0cc',
+      },
+    ]);
   });
 
   it('should clear the error when variables have changed', async () => {
@@ -166,34 +320,94 @@ describe('useQuery', () => {
       },
     });
 
-    const { result, waitForNextUpdate } = renderHook(() => useQuery(QUERY, { page: 1 }), { wrapper: StoreWrapper() });
+    let state = initialMockState;
+    const store = configureStore([])(() => state);
 
+    const { result, waitForNextUpdate } = renderHook(() => useQuery(QUERY, { page: 1 }), {
+      wrapper: StoreWrapper(store),
+    });
+
+    state = {
+      ...mockedStateAfterUsersListResolve,
+      queries: {
+        [generateQueryCacheKey({ domain: 'users', action: 'list', options: { page: 1 } })]: {
+          loading: false,
+          ids: ['e6538393-aa7e-4ec2-870b-f75b3d85f706', '708c8008-3455-49a9-b66a-5222bcadb0cc'],
+        },
+      },
+    };
     await waitForNextUpdate();
 
     act(() => {
       result.current.fetchMore({
         variables: { page: 2 },
-        updateQuery: ({ previousData, data }) => [previousData, data],
+        updateQuery: ({ previousData, data }) => [...previousData, ...data],
       });
     });
 
     expect(result.current.loading).toBeTruthy();
 
+    state = {
+      entities: {
+        users: {
+          ...state.entities.users,
+          'e57a6047-cb52-4273-9df7-1d55c2c6e36d': {
+            id: 'e57a6047-cb52-4273-9df7-1d55c2c6e36d',
+          },
+        },
+      },
+      queries: {
+        ...state.queries,
+        [generateQueryCacheKey({ domain: 'users', action: 'list', options: { page: 2 } })]: {
+          loading: false,
+          ids: ['e57a6047-cb52-4273-9df7-1d55c2c6e36d'],
+        },
+      },
+    };
     await waitForNextUpdate();
 
+    const resultData = [
+      {
+        id: 'e6538393-aa7e-4ec2-870b-f75b3d85f706',
+      },
+      {
+        id: '708c8008-3455-49a9-b66a-5222bcadb0cc',
+      },
+      { id: 'e57a6047-cb52-4273-9df7-1d55c2c6e36d' },
+    ];
+
     expect(result.current.loading).toEqual(false);
-    expect(result.current.data).toEqual(['data', 'more data']);
+    expect(result.current.data).toEqual(resultData);
   });
 
   it('should ignore the store when the ignoreCache option is passed', async () => {
     const QUERY = () => ({
-      domain: 'users',
+      domain: 'projects',
       action: 'list',
     });
 
     const storeKey = generateQueryCacheKey(QUERY());
 
-    const store = configureStore([])({ queries: { [storeKey]: 'cachedData' } });
+    const store = configureStore([])({
+      entities: {
+        projects: {
+          'e57a6047-cb52-4273-9df7-1d55c2c6e36d': {
+            'key': 'cached'
+          }
+        }
+      },
+      queries: {
+        [storeKey]: {
+          ids: ['e57a6047-cb52-4273-9df7-1d55c2c6e36d']
+        }
+      }
+    });
+
+    mockAPI.projects.list.mockReturnValueOnce(new Promise((resolve) => {
+      resolve({
+        data: []
+      })
+    }));
 
     const { result, waitForNextUpdate } = renderHook(() => useQuery(QUERY, {}, { ignoreCache: true }), {
       wrapper: StoreWrapper(store),
@@ -201,6 +415,6 @@ describe('useQuery', () => {
 
     await waitForNextUpdate();
 
-    expect(result.current.data).toEqual('data');
+    expect(mockAPI.projects.list).toHaveBeenCalled();
   });
 });