+++ src/useQuery/useQuery.ts
@@ -2,10 +2,16 @@ import { useEffect, useContext, useCallback } from 'react';
 
 import generateQueryCacheKey from '../utils/generateQueryCacheKey';
 import useUpdatableState from '../utils/useUpdatableState';
+import normalize from '../utils/normalize';
+import { Response } from '../typings/API';
 
-import { cacheQueryResult } from '../store/actions';
+import { queryRequest, querySuccess, queryFailure } from '../store/queries/actions';
 import CustomReduxContext from '../store/CustomReduxContext';
 import Context from '../Context';
+import { saveNormalizedEntities } from '../store/entities/actions';
+import { selectMergedEntities } from '../store/entities/selectors';
+import { selectQuery, selectMetaFromQuery } from '../store/queries/selectors';
+import { TYPE_DOMAIN_MAPPING } from '../store/entities/constants';
 
 type CalculatedQuery = {
   domain: string;
@@ -17,10 +23,6 @@ type Query = (variables?: any) => CalculatedQuery;
 
 type Options = {
   ignoreCache: boolean;
-};
-
-const selectQueryData = (state, key) => {
-  return state.queries[key];
 };
 
 const defaultConfig = {
@@ -36,6 +38,7 @@ const useQuery: (query: Query, variables?: any, options?: Options) => any = (
     loading: false,
     data: undefined,
     error: undefined,
+    meta: undefined,
   });
 
   const API = useContext(Context);
@@ -49,29 +52,62 @@ const useQuery: (query: Query, variables?: any, options?: Options) => any = (
   const requestData = useCallback(
     (domain, action, options, updateQuery) => {
       const key = generateQueryCacheKey({ domain, action, options });
+      const isEntityAction = action === 'info' || action === 'list';
 
       if (!ignoreCache) {
         // Check for previous results
-        const cacheResult = selectQueryData(store.getState(), key);
+        const cacheResult = selectQuery(store.getState(), key);
         if (cacheResult) {
-          setState({ loading: false, data: cacheResult, error: undefined });
+          const data = selectMergedEntities(store.getState(), { key });
+          const meta = selectMetaFromQuery(store.getState(), key);
+
+          setState({ loading: false, data, meta, error: undefined });
           return;
         }
       }
 
+      store.dispatch(queryRequest({ key }));
       setState({ loading: true, error: undefined });
 
       // @TODO this promise should be cancellable
       API[domain][action](options)
-        .then(data => {
-          store.dispatch(cacheQueryResult({ domain, action, options, data }));
+        .then((response: Response) => {
+          store.dispatch(
+            querySuccess({
+              key,
+              ...(isEntityAction && { ids: Array.isArray(response.data) ? response.data.map(entity => entity.id) : response.data.id }),
+              ...(!isEntityAction && { data: response.data }),
+              meta: response.meta,
+            }),
+          );
+
+          if (isEntityAction) {
+            const mainEntities = normalize(response.data);
+            store.dispatch(saveNormalizedEntities({ type: domain, entities: mainEntities }));
+          }
+
+          if (response.included) {
+            Object.keys(response.included).forEach(entityType => {
+              const normalizedEntities = normalize(response.included[entityType]);
+              const domainFromType = TYPE_DOMAIN_MAPPING[entityType];
+              store.dispatch(saveNormalizedEntities({ type: domainFromType, entities: normalizedEntities }));
+            });
+          }
+
+          const data = selectMergedEntities(store.getState(), { key });
+          const meta = selectMetaFromQuery(store.getState(), key);
 
           setState({
             loading: false,
             data: updateQuery ? updateQuery({ previousData: state.data, data }) : data,
+            meta,
+            error: undefined,
           });
         })
-        .catch(error => setState({ loading: false, error }));
+        .catch(error => {
+          store.dispatch(queryFailure({ key, error }));
+          setState({ loading: false, error });
+        });
     },
     [state.data],
   );